# apiEscola/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppEscola API",
        "description": "API REST do sistema multi-tenant AppEscola — gestão de escolas e cursinhos.",
        "contact": {
            "email": "contato@appescola.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Listar disciplinas",
                "operationId": "7ec04c0df2c6d2830d25036d6f4d4b61",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de disciplinas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Criar disciplina",
                "operationId": "f612fe6df780898b985b37bd4ed82da3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Dados da disciplina"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disciplina criada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subjects/{id}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Exibir disciplina",
                "operationId": "67f17334a964551021eb9c84cfc5ec9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da disciplina"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Atualizar disciplina",
                "operationId": "a2817c953cf009c853d9e90a316ecf17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Campos a atualizar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disciplina atualizada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Remover disciplina",
                "operationId": "4895aad801f749a09b54fe3a6fabb82d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido com sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Informe o token retornado pelo endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}